backend/requirements.txt
fastapi
uvicorn
motor
pydantic
scikit-learn
numpy
joblib
passlib[bcrypt]
PyJWT
python-multipart
certifi
